Fixes #34785 - Fix loading of hosts_extensions_helper
Use plugin provider methods :hosts_multiple_actions and
:host_overview_buttons for extending the host UI instead of
prepending the module to the Foreman's HostsHelper module.
